requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-generativeai
requests
beautifulsoup4
httpx
